Новости

+++ database/migrations/2023_02_02_215836_create_options_table.php
@@ -30,6 +30,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('options');
+
+        if (app()->isLocal())
+            Schema::dropIfExists('options');
     }
 };
