+++ database/migrations/2023_02_14_215944_create_products_table.php
@@ -38,6 +38,7 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('products');
+        if (app()->isLocal())
+            Schema::dropIfExists('products');
     }
 };
