+++ database/migrations/2023_02_15_124921_create_categories_table.php
@@ -36,6 +36,7 @@ return new class extends Migration
      */
     public function down()
     {
+        if (app()->isLocal())
         Schema::dropIfExists('categories');
     }
 };
