+++ database/migrations/2023_02_15_225049_create_vacancies_table.php
@@ -33,6 +33,7 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('vacancies');
+        if (app()->isLocal())
+            Schema::dropIfExists('vacancies');
     }
 };
